Fix storage setting to append hash to static files
Apparently, my Django template is out-of-date because the storage
setting to append hash to static files in order to avoid cache issues is
not the correct one. This was changed in Django 4.2, but I never noticed
the issue until now.

This will also help my other projects that were not migrated to the new
setting...